pyproject.toml
[project]
name = "reproducible-scientific-notebook-configurations"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "antimony>=3.0.0",
    "libroadrunner>=2.8.0",
    "matplotlib>=3.10.6",
    "pandas>=2.3.2",
    "python-dotenv>=1.1.1",
    "pyyaml>=6.0.2",
]
